+++ test/test_helper.rb
@@ -1,3 +1,7 @@
-require 'test/unit'
+require 'simplecov'
+
+require 'minitest/autorun'
+require 'minitest/reporters'
+Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 require File.expand_path('../../lib/edfize', __FILE__)